Implement multistep file conversion
The process accepts a mapping dict showing which suffixes can be
interconverted with which process. Networkx is used to construct a graph
containing all the processes, and to find the shortest conversion path
from start to finish.

Temporary files are used for intermediates. These are cleaned following
conversion.

Both python functions and command line calls can be used for conversion